+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "smooth-push",
-  "version": "1.5.9",
+  "version": "1.5.10",
   "author": "CK",
   "repository": {
     "type": "git",
