+++ src/consts.ts
@@ -1,0 +1,3 @@
+export const INITIAL_POSITION = -180;
+export const SWIPE_THRESHOLD = -55;
+export const DURATION = 400;
